feat(about-me): added about-me component, fonts, and mobile styling
The about me component was copied over from the old site, along with a new image. Testing was done
on responsiveness of images and font-sizes for small screens and css adjustments were made. Finally
there were slight changes to the GraphQL schema for images with alt text, and the ability to create
or delete GraphQL documents was removed to ensure singleton instances of the documents.

BREAKING CHANGE: GraphQL schema change for alt text caused previous documents to no longer be valid