feat(entire app): minimalism redesign and adding analytics
Header was restyled to remove unnecessary styling. Carousel was replaced with a much more simple
banner. Animations were simplified to reduce delay and variance, and horizontal overflow was hidden
to avoid unintended view extension. The contact section colour scheme was simplified, and given
dynamic fonts as well as moving the title to the opposite side. The footer was reorganized to look
much more cohesive. Mobile styling was improved. Analytics added for page load and external link
clicks.

fix #14, #13, #12, #11